+++ src/pkg.ts
@@ -1,4 +1,4 @@
 export const name = "presenit" as const;
-export const version = "0.7.6" as const;
+export const version = "0.7.7" as const;
 
 export { DEFAULT_CONFIG, DEFAULT_SLIDE_ALIGN } from "./types";